Changed Artifact id and fix PR comments
