+++ src/main/java/org/learntotravel/web_ui/MasterApplication.java
@@ -1,4 +1,4 @@
-package org.learntotravel.master;
+package org.learntotravel.web_ui;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
