+++ src/test/java/org/learntotravel/web_ui/MasterApplicationTests.java
@@ -1,4 +1,4 @@
-package org.learntotravel.master;
+package org.learntotravel.web_ui;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
